ISSUES-4006 split replication packets
