+++ src/Core/MySQL/PacketsReplication.h
@@ -1,0 +1,61 @@
+#pragma once
+
+#include <Core/MySQL/PacketPayloadReadBuffer.h>
+#include <Core/MySQL/PacketPayloadWriteBuffer.h>
+#include <Core/MySQL/PacketEndpoint.h>
+
+/// Implementation of MySQL wire protocol.
+/// Works only on little-endian architecture.
+
+namespace DB
+{
+
+namespace MySQLProtocol
+{
+
+namespace Replication
+{
+
+/// https://dev.mysql.com/doc/internals/en/com-register-slave.html
+class RegisterSlave : public IMySQLWritePacket
+{
+public:
+    UInt32 server_id;
+    String slaves_hostname;
+    String slaves_users;
+    String slaves_password;
+    size_t slaves_mysql_port;
+    UInt32 replication_rank;
+    UInt32 master_id;
+
+protected:
+    size_t getPayloadSize() const override;
+
+    void writePayloadImpl(WriteBuffer & buffer) const override;
+
+public:
+    RegisterSlave(UInt32 server_id_);
+};
+
+/// https://dev.mysql.com/doc/internals/en/com-binlog-dump.html
+class BinlogDump : public IMySQLWritePacket
+{
+public:
+    UInt32 binlog_pos;
+    UInt16 flags;
+    UInt32 server_id;
+    String binlog_file_name;
+
+protected:
+    size_t getPayloadSize() const override;
+
+    void writePayloadImpl(WriteBuffer & buffer) const override;
+
+public:
+    BinlogDump(UInt32 binlog_pos_, String binlog_file_name_, UInt32 server_id_);
+};
+
+}
+}
+
+}
