+++ src/Core/MySQLReplication.cpp
@@ -4,6 +4,8 @@
 #include <IO/ReadBufferFromString.h>
 #include <common/DateLUT.h>
 #include <Common/FieldVisitors.h>
+#include <Core/MySQL/PacketsGeneric.h>
+#include <Core/MySQL/PacketsProtocolText.h>
 
 namespace DB
 {
